pyproject.toml
[tool.poetry]
name = "weather-api"
version = "0.1.0"
description = "A Django REST API for weather forecasts"
authors = ["Fernando Flores <ferjflores@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "scripts:runserver"
test = "scripts:test"
test_unit = "scripts:test_unit"
test_integration = "scripts:test_integration"
makemigrations = "scripts:makemigrations"
migrate = "scripts:migrate"
showmigrations = "scripts:showmigrations"